release-metadata 1.0.0
This is the initial release of release-metadata for use with Node
applications. This is based originally on work done at Kinetic Cafe and
released as part of the Cartage gem and later ported to Elixir (not yet
open-sourced). We use this information as part of our live endpoint
test and lets us know which version of the application has been
deployed.

When writing this version, there are several changes to the original
release metadata format. There are built-in secure representations of
the release metadata format and sub-formats, and a built-in extensible
version.

Tests are currently minimal because there's a lot of data that would
need to be mocked in order to meaningfully test.